Move lua bindings into their own folder
