+++ CMakeLists.txt
@@ -210,9 +210,9 @@ set(MOMEMTA_SOURCES
     "core/src/logger/sink.cc"
     "core/src/lua/LazyTable.cc"
     "core/src/lua/ParameterSetParser.cc"
-    "core/src/lua/Path.cc"
-    "core/src/lua/Types.cc"
     "core/src/lua/utils.cc"
+    "core/src/lua/bindings/Path.cc"
+    "core/src/lua/bindings/Types.cc"
     "core/src/registry/ModuleDefBuilder.cc"
     "core/src/registry/ModuleDefUtils.cc"
     "core/src/registry/ModuleRegistry.cc"
