+++ core/src/lua/bindings/Path.cc
@@ -16,11 +16,11 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <lua/Path.h>
+#include <lua/bindings/Path.h>
 
 #include <momemta/ILuaCallback.h>
 
-#include <lua/Types.h>
+#include <lua/bindings/Types.h>
 #include <ExecutionPath.h>
 
 #include <lua.hpp>
