+++ core/src/lua/bindings/Types.cc
@@ -16,11 +16,11 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <lua/Types.h>
+#include <lua/bindings/Types.h>
 
 #include <ExecutionPath.h>
 
-#include <lua/Path.h>
+#include <lua/bindings/Path.h>
 #include <lua.hpp>
 
 void push_type_metatable(lua_State* L, const char* name) {
